requirements.txt
fastapi
uvicorn
python-multipart
Jinja2
requests
python-docx
python-jose
passlib[bcrypt]
